IP Sweeper
IP Sweep verifies syntax for a proper IPv4 net addr to scan. if an improper one is given, it prompts for a proper input. upon success, it sends a list of hosts, that are currently on the network and accept ICMP requests, to a file: $HOME/host_up.txt. the file is then displayed to the screen